Move MachineStatusChangedEvent to che-core-ide-app from ide-api
