+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/workspace/events/MachineStatusChangedEvent.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.ide.api.workspace.event;
+package org.eclipse.che.ide.workspace.events;
 
 import com.google.gwt.event.shared.EventHandler;
 import com.google.gwt.event.shared.GwtEvent;
